Small bug fixes
Now players must wait 0.5 second to be able to walk again. This procedure allows the game to adequately process the next generation of cells, which was breaking often when a player walked too fast (making successive moves in a very short period of time). Besides, I have fixed a small glitch on the config dialog.